% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoconvert.R
\name{phenoconvert}
\alias{phenoconvert}
\title{Calculates phenoconversion corrected activity scores}
\usage{
phenoconvert(input_data, id, enzyme, activity_score, medication)
}
\arguments{
\item{input_data}{data.frame with activity scores and concomitant medication}

\item{id}{individual identifier column}

\item{enzyme}{option that selects enzyme upon which phenoconversions to be applied}

\item{activity_score}{column in x that stores enzyme activity scores}

\item{medication}{column in x that stores names of concomitant medication}
}
\value{
returns dataframe with additional column containing phenoconversion corrected activity scores.
}
\description{
Given a data.frame containing enzyme activity scores and concomitant medication phenoconverteR
converts the scores using the conversion factors specified in the key.
}
\examples{
 \dontrun{
phenoconvert(input_data, "Study_ID", "CYP1A2", "CYP1A2_AS",  "other_medication")
}
}
