% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{test_data}
\alias{test_data}
\title{Function to generate a test data set that phenoconverteR can be tested on.}
\usage{
test_data()
}
\value{
a dataframe containing 100 participants with information on age, sex, weight, dose, medication, and activity scores/metabolism phenotypes for 3 CYP enzymes.
}
\description{
Function to generate a test data set that phenoconverteR can be tested on.
}
\examples{
test_df <- test_data()

}
