% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\name{test_data_multimed}
\alias{test_data_multimed}
\title{Function to generate a more complex test data set that phenoconverteR can be tested on. This has multiple medication columns, and therefore should be used with the 'transform_and_clean' function.}
\usage{
test_data_multimed()
}
\value{
a dataframe containing 100 participants with information on age, sex, weight, dose, multiple medication columns, and activity scores/metabolism phenotypes for 3 CYP enzymes.
}
\description{
Function to generate a more complex test data set that phenoconverteR can be tested on. This has multiple medication columns, and therefore should be used with the 'transform_and_clean' function.
}
\examples{
test_df <- test_data_multimed()

}
