% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoconvert.R
\name{transform_and_clean}
\alias{transform_and_clean}
\title{Function to re-code non-relevant medications and extract non-NA columns in instances where there are multiple medication columns}
\usage{
transform_and_clean(input_data, id, variable_names)
}
\arguments{
\item{input_data}{data.frame with Study ID and multiple columns specifying concomitant medication}

\item{id}{column denoting Study ID}

\item{variable_names}{list of columns in input data that store names of concomitant medication}
}
\value{
returns dataframe with cleaner containing only relevant medications, with a summary column
}
\description{
Function to re-code non-relevant medications and extract non-NA columns in instances where there are multiple medication columns
}
\examples{
\dontrun{
med_cols <- c("med1", "med2", "med3", "med4")
phenoconverter::transform_and_clean(input_data, med_cols)
}

}
